+++ packages/odata-service-inquirer/src/prompts/connectionValidator.ts
@@ -293,7 +293,7 @@ export class ConnectionValidator {
             this._validatedClient = url.searchParams.get(SAP_CLIENT_KEY) ?? undefined;
             return 200;
         } catch (e) {
-            LoggerHelper.logger.debug(`ConnectionValidator.checkSapService() - error: ${e.message}`);
+            LoggerHelper.logger.debug(`ConnectionValidator.checkUrl() - error: ${e.message}`);
             if (e?.isAxiosError) {
                 // Error handling for BAS specific 500 errors
                 if (e?.response?.status.toString().match(/5\d\d/) && isBAS) {
@@ -602,12 +602,15 @@ export class ConnectionValidator {
     ): Promise<{ valResult: ValidationResult; errorType?: ERROR_TYPE }> {
         this.resetConnectionState();
         this.resetValidity();
-        // Get the destination URL in the BAS specific form <protocol>://<destinationName>.dest
-        const destUrl = getDestinationUrlForAppStudio(destination.Name, servicePath);
+        // Get the destination URL in the BAS specific form <protocol>://<destinationName>.dest. This function lowercases the origin.
+        const destUrl = getDestinationUrlForAppStudio(destination.Name, servicePath).toLowerCase();
         // Get the destination URL in the portable form <protocol>://<host>:<port>.
         // We remove trailing slashes (up to 10, infinite would allow DOS attack) from the host to avoid double slashes when appending the service path.
         this._destinationUrl = servicePath
-            ? destUrl.replace(`https://${destination.Name}.dest`, destination.Host.replace(/\/{1,10}$/, ''))
+            ? destUrl.replace(
+                  `https://${destination.Name.toLowerCase()}.dest`,
+                  destination.Host.replace(/\/{1,10}$/, '')
+              )
             : destination.Host;
         this._destination = destination;
         // No need to apply sap-client as this happens automatically (from destination config) when going through the BAS proxy
@@ -617,7 +620,7 @@ export class ConnectionValidator {
 
         const validationResult = this.getValidationResultFromStatusCode(status);
 
-        if (this.validity.reachable && (!this.validity.authRequired || this.validity.authenticated)) {
+        if (this.validity.reachable) {
             this._validatedUrl = destUrl;
         }
 
@@ -631,7 +634,12 @@ export class ConnectionValidator {
         }
         if (this.validity.authRequired) {
             return {
-                valResult: ErrorHandler.getErrorMsgFromType(ERROR_TYPE.AUTH)!,
+                valResult: ErrorHandler.getErrorMsgFromType(
+                    ERROR_TYPE.AUTH,
+                    destination.Authentication !== Authentication.NO_AUTHENTICATION
+                        ? t('texts.checkDestinationAuthConfig')
+                        : undefined
+                )!,
                 errorType: ERROR_TYPE.AUTH
             };
         }
@@ -702,9 +710,7 @@ export class ConnectionValidator {
                 isSystem,
                 odataVersion
             });
-            LoggerHelper.logger.debug(
-                `ConnectionValidator.checkSapServiceUrl() - status: ${status}; url: ${serviceUrl}`
-            );
+            LoggerHelper.logger.debug(`ConnectionValidator.validateUrl() - status: ${status}; url: ${serviceUrl}`);
             this.validity.urlFormat = true;
             this._validatedUrl = serviceUrl;
 
@@ -903,7 +909,7 @@ export class ConnectionValidator {
                 isSystem,
                 odataVersion
             });
-            LoggerHelper.logger.debug(`ConnectionValidator.checkSapServiceUrl() - status: ${status}; url: ${url}`);
+            LoggerHelper.logger.debug(`ConnectionValidator.validateAuth() - status: ${status}; url: ${url}`);
             // Since an exception was not thrown, this is a valid url
             this.validity.urlFormat = true;
             this._validatedUrl = url;