+++ examples/fe-fpm-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fe-fpm-cli
 
+## 0.0.49
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.31.22
+
+## 0.0.48
+
+### Patch Changes
+
+-   @sap-ux/fe-fpm-writer@0.31.21
+
 ## 0.0.47
 
 ### Patch Changes