+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 0.0.5
+
+### Patch Changes
+
+-   1bb4d48: adds new module @sap-ux/abap-deploy-config-sub-generator
+-   Updated dependencies [1bb4d48]
+    -   @sap-ux/fiori-generator-shared@0.7.14
+
+## 0.0.4
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.13
+
+## 0.0.3
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.12
+
 ## 0.0.2
 
 ### Patch Changes