+++ packages/telemetry/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/telemetry
 
+## 0.5.47
+
+### Patch Changes
+
+-   Updated dependencies [e93797a]
+    -   @sap-ux/project-access@1.28.9
+
+## 0.5.46
+
+### Patch Changes
+
+-   Updated dependencies [2359524]
+    -   @sap-ux/btp-utils@0.17.1
+
 ## 0.5.45
 
 ### Patch Changes